+++ src/tr2/game/scaler.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <stdint.h>
+
+int32_t Scaler_Calc(int32_t unit);
+int32_t Scaler_CalcInverse(int32_t unit);